requirements.txt
astroid
azure-cognitiveservices-speech
certifi
chardet
click
colorama
Flask
idna
isort
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
pylint
requests
six
toml
typed-ast
urllib3
Werkzeug
wrapt